Added test coverage for betweenness_centrality algo.
