+++ ultragraph/src/traits/graph_algo_structural.rs
@@ -5,10 +5,6 @@
 use crate::{GraphError, GraphView};
 
 pub trait StructuralGraphAlgorithms<N, W>: GraphView<N, W> {
-    /// Finds all Strongly Connected Components in the graph using Tarjan's algorithm.
-    ///
-    /// # Returns
-    /// A vector of vectors, where each inner vector is a list of node indices
-    /// belonging to a single SCC.
+    /// Finds all Strongly Connected Components in the graph.
     fn strongly_connected_components(&self) -> Result<Vec<Vec<usize>>, GraphError>;
 }
