+++ ultragraph/src/traits/graph_view.rs
@@ -18,15 +18,7 @@ pub trait GraphView<N, W> {
     fn contains_edge(&self, a: usize, b: usize) -> bool;
     fn number_edges(&self) -> usize;
 
-    /// Returns a vector of immutable references to all active nodes in the graph.
-    ///
-    /// The order of the nodes in the returned vector is not guaranteed.
-    /// This operation is O(V) as it iterates through all possible node slots
     fn get_all_nodes(&self) -> Vec<&N>;
-
-    /// Retrieves a list of all outgoing edges from a given source node.
-    /// Returns `None` if the source node does not exist.
-    /// The returned vector contains tuples of `(target_node_index, edge_weight_reference)`.
     fn get_edges(&self, source: usize) -> Option<Vec<(usize, &W)>>;
 
     fn get_last_index(&self) -> Option<usize>;
