+++ ultragraph/src/utils/utils_tests.rs
@@ -1,0 +1,39 @@
+/*
+ * SPDX-License-Identifier: MIT
+ * Copyright (c) "2025" . The DeepCausality Authors and Contributors. All Rights Reserved.
+ */
+use crate::{GraphMut, UltraGraph, UltraGraphContainer};
+
+// Helper to create a graph for testing.
+// 0 -> 1 -> 2
+// ^----|    |-> 3
+// |         |
+// +---------+
+pub fn get_cyclic_graph() -> UltraGraphContainer<i32, ()> {
+    let mut g = UltraGraph::new();
+    g.add_node(0).unwrap();
+    g.add_node(1).unwrap();
+    g.add_node(2).unwrap();
+    g.add_node(3).unwrap();
+    g.add_edge(0, 1, ()).unwrap();
+    g.add_edge(1, 2, ()).unwrap();
+    g.add_edge(2, 0, ()).unwrap(); // Cycle
+    g.add_edge(2, 3, ()).unwrap();
+    g.freeze();
+    g
+}
+
+// Helper to create a DAG.
+// 0 -> 1 -> 2 -> 3
+pub fn get_acyclic_graph() -> UltraGraphContainer<i32, ()> {
+    let mut g = UltraGraph::new();
+    g.add_node(0).unwrap();
+    g.add_node(1).unwrap();
+    g.add_node(2).unwrap();
+    g.add_node(3).unwrap();
+    g.add_edge(0, 1, ()).unwrap();
+    g.add_edge(1, 2, ()).unwrap();
+    g.add_edge(2, 3, ()).unwrap();
+    g.freeze();
+    g
+}
